API lock 기능 구현 (#122)

+++ src/main/java/org/codequistify/master/domain/lab/controller/LabController.java
@@ -11,6 +11,7 @@ import org.codequistify.master.domain.stage.domain.Stage;
 import org.codequistify.master.domain.stage.service.impl.StageSearchServiceImpl;
 import org.codequistify.master.global.aspect.LogExecutionTime;
 import org.codequistify.master.global.aspect.LogMonitoring;
+import org.codequistify.master.global.lock.LockManager;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.http.HttpStatus;
@@ -21,14 +22,16 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import java.util.concurrent.locks.ReentrantLock;
+
 @RestController
 @RequiredArgsConstructor
 @Tag(name = "Lab")
 public class LabController {
     private final LabService labService;
     private final StageSearchServiceImpl stageSearchService;
+    private final LockManager lockManager;
 
-    private final String DEPLOY_HOST = "ws://ec2-13-125-76-129.ap-northeast-2.compute.amazonaws.com";
     private final String LAB_HOST = "wss://lab.pol.or.kr";
 
     private final Logger LOGGER = LoggerFactory.getLogger(LabController.class);
@@ -47,17 +50,31 @@ public class LabController {
     @PostMapping("lab/terminal/stage/{stage_id}")
     public ResponseEntity<PShellCreateResponse> applyPShell(@AuthenticationPrincipal Player player,
                                                             @PathVariable(name = "stage_id") Long stageId) {
-        Stage stage = stageSearchService.getStageById(stageId);
+        ReentrantLock lock = lockManager.getLock(player.getId(), stageId);
+        if (lock.tryLock()) {
+            try {
+                Stage stage = stageSearchService.getStageById(stageId);
 
-        labService.deleteSyncStageOnKubernetes(player, stage); // 동기 삭제
-        labService.createStageOnKubernetes(player, stage);
+                labService.deleteSyncStageOnKubernetes(player, stage); // 동기 삭제
+                labService.createStageOnKubernetes(player, stage);
 
-        PShellCreateResponse response = PShellCreateResponse
-                .of(LAB_HOST, player.getUid(), stage.getStageImage().name().toLowerCase());
+                PShellCreateResponse response = PShellCreateResponse
+                        .of(LAB_HOST, player.getUid(), stage.getStageImage().name().toLowerCase());
 
+                return ResponseEntity
+                        .status(HttpStatus.OK)
+                        .body(response);
+            } finally {
+                lockManager.unlock(player.getId(), stageId);
+            }
+        }
+
+        // 락 걸린 동안 들어오는 요청은 무시
+        LOGGER.info("[applyPShell] apply 작업중 중복된 요청 발생. stage: {}", stageId);
         return ResponseEntity
-                .status(HttpStatus.OK)
-                .body(response);
+                .status(HttpStatus.TOO_MANY_REQUESTS)
+                .body(null);
+
     }
 
     // 접속 가능한 주소 조회
